# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CUP FICCT Backend API",
        "description": "API REST del sistema de admisión al CUP FICCT.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor del backend"
        }
    ],
    "paths": {
...
        "/api/academic-management/carreras": {
            "get": {
                "tags": [
                    "Carreras"
                ],
                "summary": "Listar carreras",
                "operationId": "d4a0f35c37e7aa07fbf44e8f194d3571",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "facultad",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de carreras"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Carreras"
                ],
                "summary": "Crear carrera",
                "operationId": "b1932b35c527e22e32dc8ecb6e1c60fe",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "codigo",
                                    "nombre",
                                    "facultad_codigo"
                                ],
                                "properties": {
                                    "codigo": {
                                        "type": "string",
                                        "example": "187-09"
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "example": "Ingeniería de Sistemas"
                                    },
                                    "facultad_codigo": {
                                        "type": "string",
                                        "example": "187"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Carrera creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Carrera"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Código no coincide con la facultad"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/academic-management/carreras/{carrera}": {
            "get": {
                "tags": [
                    "Carreras"
                ],
                "summary": "Ver carrera",
                "operationId": "46dcf98b62075afe4b75c196d78d3ee0",
                "parameters": [
                    {
                        "name": "carrera",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Carrera",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Carrera"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Carreras"
                ],
                "summary": "Editar carrera",
                "operationId": "3b1af4f410992b47268a347e966de965",
                "parameters": [
                    {
                        "name": "carrera",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "codigo": {
                                        "type": "string"
                                    },
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "facultad_codigo": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Carrera actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Carrera"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Carreras"
                ],
                "summary": "Eliminar carrera",
                "operationId": "4f2cf275fe1836dadc863e877f4d47ec",
                "parameters": [
                    {
                        "name": "carrera",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Carrera eliminada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Carrera": {
                "properties": {
                    "codigo": {
                        "type": "string",
                        "example": "187-09"
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Ingeniería de Sistemas"
                    },
                    "facultad_codigo": {
                        "type": "string",
                        "example": "187"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}